Remove old itembuilder
