+++ platforms/minecraft/paper/src/main/java/com/ryderbelserion/fusion/paper/api/builders/items/types/SkullBuilder.java
@@ -1,9 +1,9 @@
-package com.ryderbelserion.fusion.paper.api.builders.items.modern.types;
+package com.ryderbelserion.fusion.paper.api.builders.items.types;
 
 import com.destroystokyo.paper.profile.PlayerProfile;
 import com.destroystokyo.paper.profile.ProfileProperty;
 import com.ryderbelserion.fusion.core.api.exceptions.FusionException;
-import com.ryderbelserion.fusion.paper.api.builders.items.modern.BaseItemBuilder;
+import com.ryderbelserion.fusion.paper.api.builders.items.BaseItemBuilder;
 import io.papermc.paper.datacomponent.DataComponentTypes;
 import io.papermc.paper.datacomponent.item.ResolvableProfile;
 import net.kyori.adventure.audience.Audience;
