+++ platforms/minecraft/paper/src/main/java/com/ryderbelserion/fusion/paper/api/builders/items/types/fireworks/FireworkStarBuilder.java
@@ -1,6 +1,6 @@
-package com.ryderbelserion.fusion.paper.api.builders.items.modern.types.fireworks;
+package com.ryderbelserion.fusion.paper.api.builders.items.types.fireworks;
 
-import com.ryderbelserion.fusion.paper.api.builders.items.modern.BaseItemBuilder;
+import com.ryderbelserion.fusion.paper.api.builders.items.BaseItemBuilder;
 import io.papermc.paper.datacomponent.DataComponentTypes;
 import org.bukkit.Color;
 import org.bukkit.FireworkEffect;
